// src/screens/withAuth/Agenda/types.ts
export interface IAgenda {
  intAgendaId: number;
  intAtendimentoId: string;
  intClienteId: string;
  strAgenda: string;
  strTelefone: string;
  strEmail: string;
  datNascimento: string;
  strCPFAgenda: string;
  strCPFCliente: string;
  intConvenioId: number;
  strConvenio: string;
  intTipoAgendamentoId: number;
  strTipoAgendamento: string;
  intProfissionalId: number;
  strProfissional: string;
  intEspecialidadeMedicaId: string;
  strEspecialidadeMedica: string;
  datAgendamento: string;
  strOrigem: string;
  intEmpresaId: number;
  strEmpresa: string;
}